up codesandbox setup script

+++ .codesandbox/setup.sh
@@ -1,8 +1,10 @@
-docker login --username adaptframework --password="$DOCKER_PASSWORD"
+[[ ! -z "$DOCKER_PASSWORD" ]] && docker login --username adaptframework --password="$DOCKER_PASSWORD"
 
-id=$(docker create adaptframework/mufl:sha-4a396e3)
-docker cp $id:/mufl /mufl
-docker rm -v $id
+if [ ! -d "/mufl" ]; then
+    id=$(docker create adaptframework/mufl:sha-4a396e3)
+    docker cp $id:/mufl /mufl
+    docker rm -v $id
+fi
 
 cd /mufl && sh build-nodejs-modules.sh
